create thoughts seed data, add user delete route

+++ controllers/user-controller.js
@@ -4,6 +4,7 @@ const userController = {
   // '/' methods
   getAllUsers(req, res) {
     User.find({})
+      .populate({ path: 'thoughts', select: '-__v'})
       .select('-__v')
       .then(dbUserData => res.json(dbUserData))
       .catch(err => {
@@ -52,15 +53,26 @@ const userController = {
         res.status(400).json(err);
       })
   },
-  deleteUser({ params }, res) {
+  async deleteUser({ params }, res) {
+    const user = await User.findOne({ _id: params.id });
+    if (!user) {
+      res.status(404).json({ message: 'No user found with this id!' });
+      return; 
+    }
+    const deletedThoughts = await Thought.deleteMany({ username: user.username });
     User.findOneAndDelete({ _id: params.id })
       .then(dbUserData => {
-        if (!dbUserData) {
-          res.status(404).json({ message: 'No user found with this id!' });
-          return;  
+        
+        const response = {
+          thoughts: deletedThoughts,
+          user: {
+            _id: dbUserData._id.toString(),
+            username: dbUserData.username,
+            email: dbUserData.email
+          }
         }
-
-        res.json(dbUserData);
+        
+        res.json(response);
       })
       .catch(err => {
         console.log(err);
